move files to relevant dirs
