+++ coast/_utils/xesmf_convert.py
@@ -1,6 +1,6 @@
 import os.path as path_lib
 import warnings
-from .gridded import Gridded
+from ..data.gridded import Gridded
 
 # from dask import delayed, compute, visualize
 # import graphviz
